# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateAccept": {
        "properties": {
          "expected_revision_id": {
            "title": "Expected Revision Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "paragraph_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph Ids"
          }
        },
        "required": [
          "expected_revision_id"
        ],
        "title": "CandidateAccept",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorAction"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WriteStoryApp backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/candidates/{candidate_id}": {
      "get": {
        "operationId": "get_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "longform"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/accept": {
      "post": {
        "operationId": "accept_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Accept Candidate",
        "tags": [
          "longform"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/reject": {
      "post": {
        "operationId": "reject_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Reject Candidate",
        "tags": [
          "longform"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/candidates": {
      "get": {
        "operationId": "list_candidates",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "longform"
        ]
      }
    },
    "/v1/chapters/{chapter_id}/findings": {
      "post": {
        "operationId": "create_user_finding",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFindingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create User Finding",
        "tags": [
          "longform"
        ]
      }
    },
    "/v1/chapters/{chapter_id}/handoff": {
      "get": {
        "operationId": "get_chapter_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Chapter Handoff",
        "tags": [
          "longform"
        ]
      },
      "put": {
        "operationId": "update_chapter_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Update Chapter Handoff",
        "tags": [
          "longform"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/plan": {
      "get": {
        "operationId": "get_chapter_plan",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Chapter Plan",
        "tags": [
          "longform"
        ]
      },
      "put": {
        "operationId": "update_chapter_plan",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Update Chapter Plan",
        "tags": [
          "longform"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/seam": {
      "get": {
        "operationId": "get_chapter_seam",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Chapter Seam",
        "tags": [
          "longform"
        ]
      }
    },
...
    "/v1/findings/{finding_id}/dismiss": {
      "post": {
        "operationId": "dismiss_finding",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Dismiss Finding",
        "tags": [
          "longform"
        ]
      }
    },
    "/v1/findings/{finding_id}/resolve": {
      "post": {
        "operationId": "resolve_finding",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Resolve Finding",
        "tags": [
          "longform"
        ]
      }
    },
...
  }
}